+++ R/ustreasuries.R
@@ -29,6 +29,7 @@
 #'            \item \bold{APY} converts Constant-Maturity Treasury (CMT) yields to
 #'     Annualized Percentage Yields (APY)
 #'     }
+#'     \item \bold{FedInvestData} Return treasury bond price data from 2010
 #'    }
 #'
 #'  \item \bold{Black Scholes Merton}